Fix query regression, update code style, add test option
The query for question 2 was optimized for speed.  This introduced a
regression which is fixed here.  This project requires PEP 8 style
conformance.  To comply with that, the code style has been updated to
limit the maximum line length to 99 and comments/docstrings to 72 (not
counting leading comments/quotes).  Finally a 'test' option was
introduced in main which exercises the other branches in the functions.